Worked on stackable item feature

+++ Assets/InventorySystem/Scripts/VicinityPackBehaviour.cs
@@ -39,31 +39,24 @@ namespace InventorySystem
                 var resultCount = Physics2D.OverlapCircleNonAlloc(PlayerObject.transform.position, radius, results, LayerMask.GetMask("Floor"));
                 if (resultCount > 0)
                 {
-                    //add found items to vicinity pack
+
                     for (int i = 0; i < resultCount; i++)
                     {
                         var itemBehaviour = results[i].GetComponent<ItemBehaviour>();
-                        //check already pickedupItems
-                        var foundSame = false;
-                        foreach (var entry in relations)
-                        {
-                            if (entry.Value.Equals(results[i].gameObject))
-                            {
-                                foundSame = true;
-                            }
-                            break;
-                        }
-                        if (!foundSame)
+                        var sameItem = GetItemById(itemBehaviour.ItemInstance.UniqueIdentifier);
+                        if (sameItem == null)
                         {
                             // make relations in case of pickups
                             var slot = AddItem(itemBehaviour);
                             relations.Add(slot.SlotIndex, results[i].gameObject);
                         }
+
                     }
                 }
             }
-
         }
+
+
 
         internal void Clear()
         {
@@ -103,19 +96,35 @@ namespace InventorySystem
             return null;
         }
 
+        public override InventorySlotBehaviour AddItem(ItemBehaviour item)
+        {
+
+            //empty slot found
+            var emptySlot = GetEmptySlot();
+            if (emptySlot == null)
+            {
+                return null;
+            }
+
+            emptySlot.AddItem(item.ItemInstance);
+
+            return emptySlot;
+        }
+
         internal void Pickup()
         {
             var actualItem = GetSelectedRelatedItem();
             if (actualItem != null)
             {
-                this.playerInventory.AddItem(actualItem);
+                var pickupResult = this.playerInventory.AddItem(actualItem);
+                if (pickupResult != null)
+                {
+                    var selectedSlot = GetSelectedSlot();
+                    selectedSlot.DeleteItem();
+                    relations.Remove(selectedSlot.SlotIndex);
+                    Destroy(actualItem.gameObject);
+                }
             }
-
-            var selectedSlot = GetSelectedSlot();
-            selectedSlot.DeleteItem();
-            relations.Remove(selectedSlot.SlotIndex);
-            Destroy(actualItem.gameObject);
-            //Debug.Log(relations);
         }
 
         internal void AddItemRelation(string slotIndex, GameObject gameObject)
